Move a function to db/data.py file
