+++ db/data.py
@@ -6,6 +6,7 @@ import json
 import os
 import db
 import db.exceptions
+from flask import current_app
 
 from sqlalchemy import text
 import sqlalchemy.exc
@@ -524,3 +525,22 @@ def get_summary_data(mbid, offset=0):
         pass
 
     return summary
+
+
+def get_new_recordings_from_lowlevel():
+    with db.engine.begin() as connection:
+        rows_to_fetch = current_app.config['RECORDINGS_FETCHED_PER_BATCH']
+
+        query = text("""SELECT lowlevel.gid
+                          FROM lowlevel
+                     LEFT JOIN musicbrainz.recording
+                            ON lowlevel.gid = musicbrainz.recording.gid
+                         WHERE musicbrainz.recording.gid is NULL
+                      ORDER BY lowlevel.id
+                         LIMIT :rows_to_fetch
+        """)
+        gids = connection.execute(query, {"rows_to_fetch": rows_to_fetch})
+        gids = gids.fetchall()
+        gids_in_AB = [value[0] for value in gids]
+
+        return gids_in_AB
