+++ musicbrainz_importer/musicbrainz_importer.py
@@ -1,6 +1,6 @@
 import logging
 import time
-import db.get_lowlevel_recordings
+import db.data
 import db.import_mb_data
 
 SLEEP_DURATION = 30  # number of seconds to wait between runs
@@ -9,7 +9,7 @@ SLEEP_DURATION = 30  # number of seconds to wait between runs
 def main():
     logging.info("musicbrainz importer started")
     while True:
-        gids_in_AB = db.get_lowlevel_recordings.get_new_recordings_from_lowlevel()
+        gids_in_AB = db.data.get_new_recordings_from_lowlevel()
         if gids_in_AB:
             logging.info("Importing MusicBrainz data...")
             logging.info('Inserting data for %d recordings...' % (len(gids_in_AB)))
